Wait for send() to complete before closing WebSocket client (#6)
I noticed that the WebSocket command was sometimes unreliable. I dug deeper and found that whilst the plugin did always connect and disconnect from the WebSocket endpoint, there were instances where the actual data did not get sent through before the disconnect. This is especially noticeable if spamming that particular Stream Deck button.

This PR bug fix resolves this issue by ensureing that send() happens before the close()

Since both send() and close() are async, there is a race condition that the send() can fail silently if the close() happened to complete first.